requirements.txt
minimalmodbus
serial
pymodbus
mysql.connector
